query_executor executes Plan instead of Statement
