+++ Cargo.toml
@@ -10,6 +10,7 @@ members = [
     "src/node",
     "src/protocol",
     "src/parser",
+    "src/plan",
     "src/query_planner",
     "src/query_executor",
     "src/sql_model",
