+++ src/query_executor/Cargo.toml
@@ -12,8 +12,8 @@ constraints = { path = "../constraints" }
 data_manager = { path = "../data_manager" }
 expr_eval = { path = "../expr_eval" }
 kernel = { path = "../kernel" }
+plan = { path = "../plan" }
 protocol = { path = "../protocol" }
-query_planner = { path = "../query_planner" }
 sql_model = { path = "../sql_model" }
 
 bigdecimal = "0.1.2"
