+++ src/query_planner/src/lib.rs
@@ -12,130 +12,86 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use std::convert::TryFrom;
+///! Module for transforming the input Query AST into representation the engine can process.
+mod create_schema;
+mod create_table;
+mod delete;
+mod drop_schema;
+mod drop_tables;
+mod insert;
+mod select;
+mod update;
 
-use sql_model::Id;
-use sqlparser::ast::ObjectName;
-use std::fmt::{self, Display, Formatter};
+use crate::{
+    create_schema::CreateSchemaPlanner, create_table::CreateTablePlanner, delete::DeletePlanner,
+    drop_schema::DropSchemaPlanner, drop_tables::DropTablesPlanner, insert::InsertPlanner, select::SelectPlanner,
+    update::UpdatePlanner,
+};
+use data_manager::DataManager;
+use plan::Plan;
+use protocol::{results::QueryError, Sender};
+use sqlparser::ast::{ObjectType, Statement};
+use std::sync::Arc;
 
-use sql_model::sql_types::SqlType;
+type Result<T> = std::result::Result<T, ()>;
 
-///! Module for representing how a query will be parameters bound, executed and
-///! values represented during runtime.
-pub mod plan;
-pub mod planner;
-
-/// represents a schema uniquely by its id
-#[derive(PartialEq, Debug, Clone)]
-pub struct SchemaId(Id);
-
-impl AsRef<Id> for SchemaId {
-    fn as_ref(&self) -> &Id {
-        &self.0
-    }
+trait Planner {
+    fn plan(self, data_manager: Arc<DataManager>, sender: Arc<dyn Sender>) -> Result<Plan>;
 }
 
-/// represents a schema uniquely by its name
-#[derive(Debug, Clone, Hash, Eq, PartialEq, Ord, PartialOrd)]
-pub struct SchemaName(String);
-
-impl AsRef<str> for SchemaName {
-    fn as_ref(&self) -> &str {
-        self.0.as_str()
-    }
+pub struct QueryPlanner {
+    data_manager: Arc<DataManager>,
+    sender: Arc<dyn Sender>,
 }
 
-impl TryFrom<&ObjectName> for SchemaName {
-    type Error = SchemaNamingError;
+impl QueryPlanner {
+    pub fn new(data_manager: Arc<DataManager>, sender: Arc<dyn Sender>) -> Self {
+        Self { data_manager, sender }
+    }
 
-    fn try_from(object: &ObjectName) -> Result<Self, Self::Error> {
-        if object.0.len() != 1 {
-            Err(SchemaNamingError(object.to_string()))
-        } else {
-            Ok(SchemaName(object.to_string()))
+    pub fn plan(&self, statement: &Statement) -> Result<Plan> {
+        match statement {
+            Statement::CreateTable { name, columns, .. } => {
+                CreateTablePlanner::new(name, columns).plan(self.data_manager.clone(), self.sender.clone())
+            }
+            Statement::CreateSchema { schema_name, .. } => {
+                CreateSchemaPlanner::new(schema_name).plan(self.data_manager.clone(), self.sender.clone())
+            }
+            Statement::Drop {
+                object_type,
+                names,
+                cascade,
+                ..
+            } => match object_type {
+                ObjectType::Table => DropTablesPlanner::new(names).plan(self.data_manager.clone(), self.sender.clone()),
+                ObjectType::Schema => {
+                    DropSchemaPlanner::new(names, *cascade).plan(self.data_manager.clone(), self.sender.clone())
+                }
+                _ => {
+                    self.sender
+                        .send(Err(QueryError::syntax_error(statement)))
+                        .expect("To Send Result to Client");
+                    Err(())
+                }
+            },
+            Statement::Insert {
+                table_name,
+                columns,
+                source,
+            } => InsertPlanner::new(table_name, columns, source).plan(self.data_manager.clone(), self.sender.clone()),
+            Statement::Update {
+                table_name,
+                assignments,
+                ..
+            } => UpdatePlanner::new(table_name, assignments).plan(self.data_manager.clone(), self.sender.clone()),
+            Statement::Delete { table_name, .. } => {
+                DeletePlanner::new(table_name).plan(self.data_manager.clone(), self.sender.clone())
+            }
+            Statement::Query(query) => {
+                SelectPlanner::new(query.clone()).plan(self.data_manager.clone(), self.sender.clone())
+            }
+            _ => Ok(Plan::NotProcessed(Box::new(statement.clone()))),
         }
-    }
-}
-
-impl Display for SchemaName {
-    fn fmt(&self, f: &mut Formatter<'_>) -> fmt::Result {
-        write!(f, "{}", self.0)
-    }
-}
-
-pub struct SchemaNamingError(String);
-
-impl Display for SchemaNamingError {
-    fn fmt(&self, f: &mut Formatter<'_>) -> fmt::Result {
-        write!(f, "only unqualified schema names are supported, '{}'", self.0)
-    }
-}
-
-/// represents a table uniquely
-#[derive(Debug, Clone, Hash, Eq, PartialEq, Ord, PartialOrd)]
-pub struct FullTableName(SchemaName, String);
-
-impl FullTableName {
-    pub fn as_tuple(&self) -> (&str, &str) {
-        (&self.0.as_ref(), &self.1)
-    }
-}
-
-impl Display for FullTableName {
-    fn fmt(&self, f: &mut Formatter<'_>) -> fmt::Result {
-        write!(f, "{}.{}", self.0.as_ref(), self.1)
-    }
-}
-
-impl TryFrom<&ObjectName> for FullTableName {
-    type Error = TableNamingError;
-
-    fn try_from(object: &ObjectName) -> Result<Self, Self::Error> {
-        if object.0.len() == 1 {
-            Err(TableNamingError::Unqualified(object.to_string()))
-        } else if object.0.len() != 2 {
-            Err(TableNamingError::NotProcessed(object.to_string()))
-        } else {
-            let table_name = object.0.last().unwrap().value.clone();
-            let schema_name = object.0.first().unwrap().value.clone();
-            Ok(FullTableName(SchemaName(schema_name), table_name))
-        }
-    }
-}
-
-pub enum TableNamingError {
-    Unqualified(String),
-    NotProcessed(String),
-}
-
-impl Display for TableNamingError {
-    fn fmt(&self, f: &mut Formatter<'_>) -> fmt::Result {
-        match self {
-            TableNamingError::Unqualified(table_name) => write!(
-                f,
-                "unsupported table name '{}'. All table names must be qualified",
-                table_name
-            ),
-            TableNamingError::NotProcessed(table_name) => write!(f, "unable to process table name '{}'", table_name),
-        }
-    }
-}
-
-/// A type of a column
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct ColumnType {
-    #[allow(dead_code)]
-    nullable: bool,
-    sql_type: SqlType,
-}
-
-/// represents a table uniquely
-#[derive(Debug, Clone, Hash, Eq, PartialEq, Ord, PartialOrd)]
-pub struct TableId((Id, Id));
-
-impl AsRef<(Id, Id)> for TableId {
-    fn as_ref(&self) -> &(Id, Id) {
-        &self.0
     }
 }
